add fake element

+++ database/migrations/2023_05_25_123559_create_trains_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('Departure_time');
             $table->string('Arrival_time');
             $table->smallInteger('Train_Code') ;
-            $table->tinyInteger('Number_of_carriages');
+            $table->smallInteger('Number_of_carriages');
             $table->boolean('In_time') ;
             $table->boolean('Deleted');
         });
